Begin classes for 3D animation of circuit.
Design:
Pressing "Animate Model" seta two flags, animating and newAnimation.
If both flags are set, the draw loop instantiates an object of type Animation.
newAnimation is then set to false. While animating is true, the draw
loop calls displayAnimation(). This in turn
causes the Towers (at the junctions) and the Walls (for each component)
to draw themselves. It also makes a set of balls start moving on
the tops of the Walls at speeds proportional to the currents in
the corresponding components.

In this version, the overall class structure is laid out and
the Tower class is implemented to display the Towers.