Fix all known bugs.
- Fixed errors in Circuit class' CalculatePotentials(). Instead of asking
components for their voltages, voltage-drops are calculated on the fly
within CalculatePotentials(). This makes more sense, since the sign of a
voltage drop of a component is not well-defined without knowing which
end you are standing on. Also fixed the for-loop: put "break" in the
right place, etc.

- Removed the "voltage" instance variables from the Component class.
Replaced it with a battVoltage field in the Battery class. battVoltage
is a positive-definite quantity.

- Added an equals() method to the Component class. (or maybe I did
that in the previous commit...) Updated Component comparisons to
.equals() instead of ==.