Make Ball spacing the same in slopes and flat wires.
- Each Wall now keeps track of the maxBalls it can have, since this can be
greater than Animation.BALLS_PER_WALL (which now is the number of balls
on a flat wall, i.e. a wire).

- Adjusted the speeds of the balls in the Ball's move() method, so that a Wall
with more Balls has the Balls moving at slower speeds (to represent constant
current).

- Changed Ball's x position from a float to a double. May not have been
necessary, but it made the code a bit cleaner. This can be changed back for
a possible performance gain (though I doubt it...)

- Bug Fix: When addNewBall() was called within Animation, and more importantly
within Wall's updeateBalls(), the x-value of the new Ball was truncated to an
int, instead of maintaining the full double precision. This caused new Balls
to be placed to close to the upstream Tower and thus they would reach the
downstream Tower too late. Eventually this sometimes resulted in piling up of
Balls on the upstream Tower, and subsequent components developed large gaps
on their upstream ends.

- Code clean-up: Within Wall, eliminated the variable numBalls, since it was
identical to balls.size()